Refactor MatrixInput and matrixmatch
