+++ src/lib/components/ViewFeedback.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import * as Card from '$lib/components/ui/card/index.js';
-	import Check from '@lucide/svelte/icons/check';
 	import ArrowLeft from '@lucide/svelte/icons/arrow-left';
 	import {
 		buildQuestionSetGroups,
@@ -9,23 +8,17 @@
 	} from '$lib/helpers/questionSetHelpers';
 	import { t } from 'svelte-i18n';
 	import { question_type_enum, type TCandidate, type TSelection } from '$lib/types';
-	import {
-		getQuestionResult,
-		GRADABLE_QUESTION_TYPES,
-		parseMatrixAnswer,
-		getMatrixCellStatus
-	} from '$lib/helpers/feedbackHelpers';
-	import { parseJsonRecord } from '$lib/helpers/matrixHelpers';
+	import { getQuestionResult, GRADABLE_QUESTION_TYPES } from '$lib/helpers/feedbackHelpers';
 	import RichText from './RichText.svelte';
-	import type { TMatrixOptions, TMatrixInputOptions } from '$lib/types';
 	import QuestionMedia from './QuestionMedia.svelte';
 	import ResultBadge from './ResultBadge.svelte';
-	import { cn } from '$lib/utils';
 	import { navState } from '$lib/navState.svelte';
 	import ChoiceAnswer from './answer/ChoiceAnswer.svelte';
 	import SubjectiveAnswer from './answer/SubjectiveAnswer.svelte';
 	import NumericalAnswer from './answer/NumericalAnswer.svelte';
 	import MatrixRatingAnswer from './answer/MatrixRatingAnswer.svelte';
+	import MatrixMatchAnswer from './answer/MatrixMatchAnswer.svelte';
+	import MatrixInputAnswer from './answer/MatrixInputAnswer.svelte';
 
 	let {
 		candidate,
@@ -121,126 +114,19 @@
 						variant="card"
 					/>
 				{:else if item.question.question_type === question_type_enum.MATRIXINPUT}
-					{@const matrixOpts = item.question.options as unknown as TMatrixInputOptions}
-					{@const inputType = matrixOpts.columns.input_type}
-					{@const matrixInputAnswer = parseJsonRecord<string>(item.fb.submitted_answer)}
-					<div class="overflow-x-auto">
-						<div class="border-border overflow-hidden rounded-xl border">
-							<table class="w-full border-collapse text-sm">
-								<thead>
-									<tr class="border-border bg-muted border-b">
-										<th class="text-foreground px-4 py-3 text-left font-semibold">
-											{matrixOpts.rows.label}
-										</th>
-										<th class="text-foreground px-4 py-3 text-left font-semibold">
-											{matrixOpts.columns.label}
-										</th>
-									</tr>
-								</thead>
-								<tbody>
-									{#each matrixOpts.rows.items as row (row.id)}
-										<tr class="border-border border-b last:border-b-0">
-											<td class="w-1/2 px-4 py-3 font-medium">{row.value}</td>
-											<td class="w-1/2 px-4 py-3">
-												<input
-													type={inputType}
-													readonly
-													class="border-input bg-background w-full rounded-lg border px-3 py-2 text-sm"
-													value={matrixInputAnswer[String(row.id)] ?? ''}
-												/>
-											</td>
-										</tr>
-									{/each}
-								</tbody>
-							</table>
-						</div>
-					</div>
+					<MatrixInputAnswer
+						question={item.question}
+						{candidate}
+						selections={toAnswerSelections(item)}
+						variant="card"
+					/>
 				{:else if item.question.question_type === question_type_enum.MATRIXMATCH}
-					{@const matrix = item.question.options as TMatrixOptions}
-					{@const matrixRows = matrix.rows.items}
-					{@const matrixColumns = matrix.columns.items}
-					{@const submittedMatrix = parseMatrixAnswer(item.fb.submitted_answer)}
-					{@const correctMatrix = parseMatrixAnswer(item.fb.correct_answer)}
-
-					<div class="mb-4 grid grid-cols-1 gap-4 md:grid-cols-2">
-						<div class="border-border overflow-hidden rounded-xl border">
-							<div
-								class="bg-muted text-foreground py-3 text-center text-xs font-bold tracking-widest uppercase"
-							>
-								{matrix.rows.label}
-							</div>
-							{#each matrixRows as row (row.id)}
-								<div class="border-border flex items-start gap-3 border-t px-4 py-3">
-									<span class="text-foreground min-w-4 shrink-0 text-sm font-bold">{row.key}</span>
-									<span class="text-foreground text-sm">{row.value}</span>
-								</div>
-							{/each}
-						</div>
-						<div class="border-border overflow-hidden rounded-xl border">
-							<div
-								class="bg-muted text-foreground py-3 text-center text-xs font-bold tracking-widest uppercase"
-							>
-								{matrix.columns.label}
-							</div>
-							{#each matrixColumns as col (col.id)}
-								<div class="border-border flex items-start gap-3 border-t px-4 py-3">
-									<span class="text-foreground min-w-4 shrink-0 text-sm font-bold">{col.key}</span>
-									<span class="text-foreground text-sm">{col.value}</span>
-								</div>
-							{/each}
-						</div>
-					</div>
-
-					<div class="overflow-x-auto">
-						<table class="w-full border-collapse text-sm">
-							<thead>
-								<tr class="bg-muted">
-									<th class="border-border w-14 border px-4 py-3"></th>
-									{#each matrixColumns as col (col.id)}
-										<th
-											class="border-border text-foreground min-w-16 border px-4 py-3 text-center font-semibold"
-										>
-											{col.key}
-										</th>
-									{/each}
-								</tr>
-							</thead>
-							<tbody>
-								{#each matrixRows as row (row.id)}
-									<tr class="border-border border-t">
-										<td
-											class="border-border text-foreground border px-4 py-3 text-center text-sm font-semibold"
-										>
-											{row.key}
-										</td>
-										{#each matrixColumns as col (col.id)}
-											{@const status = getMatrixCellStatus(
-												row.id,
-												col.id,
-												submittedMatrix,
-												correctMatrix
-											)}
-											<td class="border-border border px-4 py-3 text-center">
-												<div
-													class={cn(
-														'mx-auto flex h-5 w-5 items-center justify-center rounded border-2',
-														status === 'correct' && 'bg-success border-success',
-														status === 'missed' && 'bg-card border-success',
-														status === 'wrong' && 'bg-error border-error',
-														status === 'none' && 'bg-card border-border'
-													)}
-												>
-													{#if status === 'correct' || status === 'wrong'}
-														<Check size={14} class="text-primary-foreground" />
-													{/if}
-												</div>
-											</td>
-										{/each}
-									</tr>
-								{/each}
-							</tbody>
-						</table>
-					</div>
+					<MatrixMatchAnswer
+						question={item.question}
+						{candidate}
+						selections={toAnswerSelections(item)}
+						variant="card"
+					/>
 				{:else if item.question.question_type === question_type_enum.SUBJECTIVE}
 					<SubjectiveAnswer
 						question={item.question}
